Move blob supplier to correct dir
