+++ ethereum/spec/src/property-test/java/tech/pegasys/teku/spec/logic/versions/deneb/helpers/MiscHelpersDenebPropertyTest.java
@@ -27,8 +27,8 @@ import tech.pegasys.teku.spec.config.SpecConfigDeneb;
 import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlobsSidecar;
 import tech.pegasys.teku.spec.datastructures.execution.Transaction;
 import tech.pegasys.teku.spec.propertytest.suppliers.SpecSupplier;
+import tech.pegasys.teku.spec.propertytest.suppliers.blobs.versions.deneb.BlobsSidecarSupplier;
 import tech.pegasys.teku.spec.propertytest.suppliers.execution.TransactionSupplier;
-import tech.pegasys.teku.spec.propertytest.suppliers.execution.versions.deneb.BlobsSidecarSupplier;
 import tech.pegasys.teku.spec.propertytest.suppliers.type.Bytes32Supplier;
 import tech.pegasys.teku.spec.propertytest.suppliers.type.KZGCommitmentSupplier;
 import tech.pegasys.teku.spec.propertytest.suppliers.type.UInt64Supplier;
