+++ ethereum/spec/src/testFixtures/java/tech/pegasys/teku/spec/propertytest/suppliers/blobs/versions/deneb/BlindedBlobSidecarSupplier.java
@@ -11,16 +11,16 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.propertytest.suppliers.blocks.versions.deneb;
+package tech.pegasys.teku.spec.propertytest.suppliers.blobs.versions.deneb;
 
 import tech.pegasys.teku.spec.SpecMilestone;
-import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlobSidecars;
+import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlindedBlobSidecar;
 import tech.pegasys.teku.spec.propertytest.suppliers.DataStructureUtilSupplier;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
 
-public class BlobSidecarsSupplier extends DataStructureUtilSupplier<BlobSidecars> {
+public class BlindedBlobSidecarSupplier extends DataStructureUtilSupplier<BlindedBlobSidecar> {
 
-  public BlobSidecarsSupplier() {
-    super(DataStructureUtil::randomBlobSidecars, SpecMilestone.DENEB);
+  public BlindedBlobSidecarSupplier() {
+    super(DataStructureUtil::randomBlindedBlobSidecar, SpecMilestone.DENEB);
   }
 }
