+++ ethereum/spec/src/testFixtures/java/tech/pegasys/teku/spec/propertytest/suppliers/blobs/versions/deneb/SignedBlindedBlobSidecarSupplier.java
@@ -11,16 +11,17 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.propertytest.suppliers.blocks.versions.deneb;
+package tech.pegasys.teku.spec.propertytest.suppliers.blobs.versions.deneb;
 
 import tech.pegasys.teku.spec.SpecMilestone;
-import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlindedBlobSidecars;
+import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.SignedBlindedBlobSidecar;
 import tech.pegasys.teku.spec.propertytest.suppliers.DataStructureUtilSupplier;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
 
-public class BlindedBlobSidecarsSupplier extends DataStructureUtilSupplier<BlindedBlobSidecars> {
+public class SignedBlindedBlobSidecarSupplier
+    extends DataStructureUtilSupplier<SignedBlindedBlobSidecar> {
 
-  public BlindedBlobSidecarsSupplier() {
-    super(DataStructureUtil::randomBlindedBlobSidecars, SpecMilestone.DENEB);
+  public SignedBlindedBlobSidecarSupplier() {
+    super(DataStructureUtil::randomSignedBlindedBlobSidecar, SpecMilestone.DENEB);
   }
 }
